Distance Changing + Item Out Of Bounds
Distance Changing: If too close to a object then the item held will move closer to the player. If the item is too close to the player then it will drop.

Item Out Of Bounds: If an item falls into a out of bounds area then it will reset to its start location.